fix: not editable 불러오기
