+++ app/structure/lib/config/user_router.dart
@@ -9,7 +9,7 @@ import 'package:provider/provider.dart';
 import 'package:structure/model/meat_model.dart';
 import 'package:structure/model/user_model.dart';
 import 'package:structure/screen/data_management/normal/edit_meat_data_screen.dart';
-import 'package:structure/screen/data_management/normal/not_editable/freshmeat_eval_not_editable_sceen.dart';
+import 'package:structure/screen/data_management/normal/not_editable/sensory_eval_not_editable_sceen.dart';
 import 'package:structure/screen/data_management/normal/not_editable/insertion_meat_image_not_editable_screen.dart';
 import 'package:structure/screen/data_management/normal/not_editable/insertion_meat_info_not_editable_screen.dart';
 import 'package:structure/screen/data_management/researcher/add_processed_meat_main_screen.dart';
@@ -43,7 +43,7 @@ import 'package:structure/viewModel/data_management/researcher/add_raw_meat_view
 import 'package:structure/viewModel/data_management/researcher/data_add_home_view_model.dart';
 import 'package:structure/viewModel/data_management/normal/data_management_view_model.dart';
 import 'package:structure/viewModel/data_management/normal/edit_meat_data_view_model.dart';
-import 'package:structure/viewModel/data_management/normal/not_editable/freshmeat_eval_not_editable_view_model.dart';
+import 'package:structure/viewModel/data_management/normal/not_editable/sensoty_eval_not_editable_view_model.dart';
 import 'package:structure/viewModel/data_management/normal/not_editable/insertion_meat_image_not_editable_view_model.dart';
 import 'package:structure/viewModel/data_management/normal/not_editable/insertion_meat_info_not_editable_view_model.dart';
 import 'package:structure/viewModel/data_management/researcher/data_management_home_tab_view_model.dart';
@@ -263,8 +263,7 @@ class UserRouter {
                       path: 'image',
                       builder: (context, state) => ChangeNotifierProvider(
                         create: (context) =>
-                            InsertionMeatImageNotEditableViewModel(
-                                meatModel, userModel, 0),
+                            InsertionMeatImageNotEditableViewModel(meatModel),
                         child: const InsertionMeatImageNotEditableScreen(),
                       ),
                     ),
@@ -272,8 +271,8 @@ class UserRouter {
                       path: 'freshmeat',
                       builder: (context, state) => ChangeNotifierProvider(
                         create: (context) =>
-                            FreshMeatEvalNotEditableViewModel(meatModel, false),
-                        child: const FreshMeatEvalNotEditableScreen(),
+                            SensoryEvalNotEditableViewModel(meatModel, false),
+                        child: const SensoryEvalNotEditableScreen(),
                       ),
                     ),
                     // 수정 가능
@@ -374,7 +373,7 @@ class UserRouter {
                           builder: (context, state) => ChangeNotifierProvider(
                             create: (context) =>
                                 InsertionMeatImageNotEditableViewModel(
-                                    meatModel, userModel, 0),
+                                    meatModel),
                             child: const InsertionMeatImageNotEditableScreen(),
                           ),
                         ),
@@ -382,9 +381,9 @@ class UserRouter {
                           path: 'freshmeat',
                           builder: (context, state) => ChangeNotifierProvider(
                             create: (context) =>
-                                FreshMeatEvalNotEditableViewModel(
+                                SensoryEvalNotEditableViewModel(
                                     meatModel, false),
-                            child: const FreshMeatEvalNotEditableScreen(),
+                            child: const SensoryEvalNotEditableScreen(),
                           ),
                         ),
                       ],
@@ -461,8 +460,7 @@ class UserRouter {
                       path: 'image',
                       builder: (context, state) => ChangeNotifierProvider(
                         create: (context) =>
-                            InsertionMeatImageNotEditableViewModel(
-                                meatModel, userModel, 0),
+                            InsertionMeatImageNotEditableViewModel(meatModel),
                         child: const InsertionMeatImageNotEditableScreen(),
                       ),
                     ),
@@ -470,8 +468,8 @@ class UserRouter {
                       path: 'freshmeat',
                       builder: (context, state) => ChangeNotifierProvider(
                         create: (context) =>
-                            FreshMeatEvalNotEditableViewModel(meatModel, false),
-                        child: const FreshMeatEvalNotEditableScreen(),
+                            SensoryEvalNotEditableViewModel(meatModel, false),
+                        child: const SensoryEvalNotEditableScreen(),
                       ),
                     ),
                   ],
